\name{BIOdry-package}
\alias{BIOdry-package}
\alias{BIOdry}
\docType{package}
\title{Multilevel Modeling of Dendroclimatical Fluctuations}

\description{{Multilevel data frames of tree rings and climate are
modeled into normalized fluctuations of tree growth and aridity.
Modeled fluctuations (model frames) are compared with Mantel
correlograms on multiple levels defined by sample design. Package
implementation can be understood by running examples in functions
ringApply(), modelFrame(), and muleMan().}  }

\details{

The DESCRIPTION file:
\packageDESCRIPTION{BIOdry}
\packageIndices{BIOdry}
}
\author{{Wilson Lara <wilarhen@gmail.com>, Felipe Bravo <fbravo@pvs.uva.es>}

Maintainer: \packageMaintainer{BIOdry}
}

\keyword{ package }


