\name{Pradii03}
\docType{data}
\alias{Pradii03}
\title{Reference inside-bark radii vector
}

\description{Numeric vector of inside-bark radii that were measured in
  dominant trees of pine forests (Pinus pinaster Ait.) in 2003. The pine
  forests are located in northern (Soria) and east-central Spain
  (Cuenca). }

\usage{data(Pradii03)}
\format{
numeric vector with names of the vector indicating the measured tree.
}

\details{This vector contains inside-bark radii, measured in four trees
during 2003 (mm). This radial vector was derived from the over-bark
diameters at breast height (cm), with allometric estimation of
inside-bark diameters from the over-bark diameters. Names of the vector
correspond to labels at tree level }

\source{Sustainable Forest Management Research Institute, University of
  Valladolid (UVa)-INIA, Spain.
}
\examples{
str(Pradii03)
}
\keyword{datasets}
