\name{Prec}
\docType{data}
\alias{Prec}
\title{Cumulative precipitations   
}
\description{Monthly cumulative precipitation (mm). 
}
\usage{data(Prec)}
\format{\code{data.frame} object with 24 records.
    \describe{
      \item{Column}{Column names contain plot codes and monthly
	abbreviations}
      }
}

\details{Recorded months begin at January and
end at December, and observed years span from 1951 to
2005. Factor in data set defines one sample level (plot)
}
\references{
  Bogino, S., and Bravo, F. (2008). Growth response of Pinus pinaster Ait. to climatic variables in central Spanish forests. Ann. For. Sci. 65, 1-13.
}
\examples{
str(Prec)
}
\keyword{data sets}
