\name{BIOdry-package}
\alias{BIOdry-package}
\alias{BIOdry}
\docType{package}
\title{Multilevel Modeling of Dendroclimatical Fluctuations}

\description{Multilevel ecological data series (MEDS) are sequences of observations ordered according to temporal/spatial hierarchies that are defined by sample designs, with sample variability confined to ecological factors. Dendroclimatic MEDS of tree rings and climate are modeled into normalized fluctuations of tree growth and aridity.  Modeled fluctuations (model frames) are compared with Mantel correlograms on multiple levels defined by sample design. Package implementation can be understood by running examples in modelFrame(), and muleMan() functions.}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{BIOdry}
\packageIndices{BIOdry}
Maintainer: \packageMaintainer{BIOdry}
}
\author{Wilson Lara <wilarhen@gmail.com>, Felipe Bravo <fbravo@pvs.uva.es>}



\keyword{ package }


