\name{Pradii03}
\docType{data}
\alias{Pradii03}
\title{Vector of radii measured on Pinus pinaster
}
\description{Vector of radii measured in dominant trees of Pinus pinaster
  (2003) growing on the North of Spain (Soria) and on the central
  portion of the same country (Cuenca)  
}
\usage{data(Pradii03)}
\format{
A numeric vector with names of the vector indicating the measured tree.
}
\details{This vector contains inside-bark radii, measured in four trees during 2003 (mm). This radial vector was derived from the over-bark
diameters at breast height (cm), with allometric estimation
of inside-bark diameters from the over-bark diameters. Names of the
vector correspond to labels at tree level
}
\source{Sustainable Forest Management Research Institute, University of
  Valladolid (UVa)-INIA, Spain.
}
\examples{
str(Pradii03)
}
\keyword{datasets}
