\name{Pradii03}
\docType{data}
\alias{Pradii03}
\title{Vector of radii measured on P.pinaster
}
\description{Vector of radii measured in dominant trees of P.pinaster
  (2003) growing on the Ebro river basin, Spain 
}
\usage{data(Pradii03)}
\format{
A numeric vector with names of the vector indicating the measured tree.
}
\details{This vector contains inside-bark radii, measured in four trees during 2003 (mm). This radial vector was derived from the over-bark
diameters at breast height (cm), with allometric estimation
of inside-bark diameters from the over-bark diameters. Names of the
vector correspond to labels at tree level
}
\source{Sustainable Forest Management Research Institute, Universidad de
  Valladolid (UVa)-INIA.
}
\examples{
str(Pradii03)
}
\keyword{datasets}
